Stealthily allow setting the volume above 65536. Implements #7
This implements feature request #7 in the laziest way possible.
If you set volume to "max" it will still be at 65535, but the volume
slider can now be set to any number Inform understands (which seems to
be capped at the Mersenne prime 2147483647.)

I won't be surprised if someone reports this as a bug on day, unless
I document it somehow, but this will have to do for now.